Rename namesapce file
