+++ plugins/modules/module_utils/mongodb_common.py
